+++ pipelines/demo/prediction/assets/tfdv_schema_serving.pbtxt
