+++ pipelines/demo/training/queries/ingest.sql
