+++ pipelines/demo/training/queries/sample.sql
