--- pipelines/tensorflow/prediction/payloads/prod.json
@@ -1,23 +1,0 @@
-{
-    "attributes": { 
-        "enable_caching": "False",
-        "template_path": "gs://<GCS path where files are published - matches _PIPELINE_PUBLISH_GCS_PATH >/<Git tag>/prediction/prediction.json"
-    },
-    "data": {
-        "project_id": "my-prod-gcp-project",
-        "project_location": "europe-west4",
-        "pipeline_files_gcs_path": "gs://<GCS path where files are published - matches _PIPELINE_PUBLISH_GCS_PATH >/<Git tag>/prediction/files",
-        "ingestion_project_id": "my-prod-gcp-project",
-        "model_name": "tensorflow-with-preprocessing",
-        "model_label": "label_name",
-        "tfdv_schema_filename": "tfdv_schema_serving.pbtxt",
-        "tfdv_train_stats_path": "gs://my-prod-bucket/train_stats/train.stats",
-        "dataset_id": "preprocessing_prediction",
-        "dataset_location": "EU",
-        "ingestion_dataset_id": "chicago_taxi_trips",
-        "timestamp":"2021-08-01 00:00:00",
-        "batch_prediction_machine_type": "n1-standard-4",
-        "batch_prediction_min_replicas": 3,
-        "batch_prediction_max_replicas": 10
-    }
-}
