--- pipelines/tensorflow/training/payloads/dev.json
@@ -1,20 +1,0 @@
-{
-    "attributes": { 
-        "enable_caching": "False",
-        "template_path": "training.json"
-    },
-    "data": {
-        "project_id": "my-sandbox-gcp-project",
-        "project_location": "europe-west4",
-        "pipeline_files_gcs_path": "gs://my-sandbox-bucket/pipelines/my-unique-namespace",
-        "ingestion_project_id": "my-sandbox-gcp-project",
-        "model_name": "tensorflow-with-preprocessing",
-        "model_label": "label_name",
-        "tfdv_schema_filename": "tfdv_schema_training.pbtxt",
-        "tfdv_train_stats_path": "gs://my-sandbox-bucket/train_stats/my-unique-namespace/train.stats",
-        "dataset_id": "preprocessing",
-        "dataset_location": "EU",
-        "ingestion_dataset_id": "chicago_taxi_trips",
-        "timestamp": "2021-08-01 00:00:00"
-    }
-}
