--- pipelines/xgboost/training/queries/engineer_features.sql
@@ -1,4 +1,0 @@
-/* The Purpose of this query is to clean the data before the training*/
-SELECT
- *
-FROM `{{ source_dataset }}.{{ source_table }}`
